--- cli-client/main.go
@@ -1,9 +1,0 @@
-package main
-
-import (
-	cli "github.com/Mahmoud-Emad/jimber/cli/commands"
-)
-
-func main() {
-	cli.InitCommands()
-}
