+++ client/models.go
@@ -1,0 +1,7 @@
+package client
+
+import cli "jimber.com/sdk/cmd/cli/commands"
+
+type JimberClient struct {
+	Commands cli.JimberCommands
+}
