+++ cmd/cli-client/main.go
@@ -1,0 +1,9 @@
+package main
+
+import (
+	client "jimber.com/sdk/client"
+)
+
+func main() {
+	client.NewJimberClient()
+}
