+++ cmd/cli/commands/models.go
@@ -1,0 +1,7 @@
+package cmd
+
+import cobra "github.com/spf13/cobra"
+
+type JimberCommands struct {
+	command *cobra.Command
+}
