+++ cmd/cmd.go
@@ -1,0 +1,7 @@
+package cmd
+
+type JimberCLI struct{}
+
+func NewJimberCLI() *JimberCLI {
+	return &JimberCLI{}
+}
