+++ cmd/logger/models.go
@@ -1,0 +1,7 @@
+package cmd
+
+type Logger struct {
+	Text   string
+	Status int
+	Color  string
+}
