--- database/queries/check_projects_table.sql
@@ -1,6 +1,0 @@
-SELECT EXISTS (
-    SELECT 1
-    FROM   information_schema.tables
-    WHERE  table_schema = 'public'
-    AND    table_name = 'projects'
-)
