--- database/queries/create_tables.sql
@@ -1,14 +1,0 @@
-CREATE TABLE projects (
-    id SERIAL PRIMARY KEY,
-    name VARCHAR(255) NOT NULL,
-    token VARCHAR(255) NOT NULL,
-    created_date TIMESTAMP NOT NULL,
-    updated_date TIMESTAMP NOT NULL,
-    owner_id INT NOT NULL
-);
-
-CREATE TABLE users (
-    id SERIAL PRIMARY KEY,
-    username VARCHAR(255) NOT NULL,
-    password VARCHAR(255) NOT NULL
-);
