+++ go.mod
@@ -1,15 +1,8 @@
-module github.com/Mahmoud-Emad/jimber
+module jimber.com/sdk
 
 go 1.20
 
-require (
-	github.com/fatih/color v1.15.0
-	github.com/gin-gonic/gin v1.9.0
-	github.com/go-ini/ini v1.67.0
-	github.com/lib/pq v1.10.9
-	github.com/spf13/cobra v1.7.0
-	golang.org/x/crypto v0.9.0
-)
+require gorm.io/driver/postgres v1.5.2
 
 require (
 	github.com/bytedance/sonic v1.8.0 // indirect
@@ -32,10 +25,24 @@ require (
 	github.com/twitchyliquid64/golang-asm v0.15.1 // indirect
 	github.com/ugorji/go/codec v1.2.9 // indirect
 	golang.org/x/arch v0.0.0-20210923205945-b76863e36670 // indirect
-	golang.org/x/net v0.10.0 // indirect
-	golang.org/x/sys v0.8.0 // indirect
-	golang.org/x/term v0.8.0 // indirect
-	golang.org/x/text v0.9.0 // indirect
+	golang.org/x/net v0.9.0 // indirect
+	golang.org/x/sys v0.7.0 // indirect
+	golang.org/x/term v0.7.0 // indirect
 	google.golang.org/protobuf v1.28.1 // indirect
 	gopkg.in/yaml.v3 v3.0.1 // indirect
+)
+
+require (
+	github.com/fatih/color v1.15.0
+	github.com/gin-gonic/gin v1.9.0
+	github.com/go-ini/ini v1.67.0
+	github.com/jackc/pgpassfile v1.0.0 // indirect
+	github.com/jackc/pgservicefile v0.0.0-20221227161230-091c0ba34f0a // indirect
+	github.com/jackc/pgx/v5 v5.3.1 // indirect
+	github.com/jinzhu/inflection v1.0.0 // indirect
+	github.com/jinzhu/now v1.1.5 // indirect
+	github.com/spf13/cobra v1.7.0
+	golang.org/x/crypto v0.8.0
+	golang.org/x/text v0.9.0 // indirect
+	gorm.io/gorm v1.25.0
 )
