--- main.go
@@ -1,13 +1,0 @@
-package main
-
-import (
-	"github.com/Mahmoud-Emad/jimber/server"
-)
-
-func main() {
-	server := server.Server{
-		Port: "8080",
-		Host: "localhost",
-	}
-	server.Serve()
-}
