+++ server/routes.go
@@ -1,0 +1,13 @@
+package server
+
+// registerAPIRoutes registers the API endpoints.
+func (srv *JimberServer) registerAPIRoutes() {
+	// api := srv.Router.Group("/api")
+	// srv.projectRoutes(api)
+}
+
+// Register all project endpoints
+// func (srv *JimberServer) projectRoutes(api *gin.RouterGroup) {
+// 	api.GET("/projects", srv.getProjectsHandler)
+// 	api.POST("/projects", srv.createProjectHandler)
+// }
